Files from custom work while learning
I originally cloned from GItHub Code: PacktPublishing / Modular-Programming-with-Python directly to PyCharm. This resulted in only being able to merge back to GItHub Code: PacktPublishing / Modular-Programming-with-Python. This is not what I desired to do because I am not intending to generate any pull requests back to the original project.

So, I started over.

I forked from GitHub Code: PacktPublishing / Modular-Programming-with-Python to Fork to: Carl-Ty/Modular-Programming-with-Python.

Then I used GitHub desktop to create a local repository configured for both in origin/master and an upstream/master.

Then I copied all my custom work from a zip file that I had saved into the new local repository

This provided the ability to push back to the Fort version of the repository. As well as, the option to eventually push back to the upstream version of the repository. That is the Packt repository.